Eliminar vendedores

+++ admin/index.php
@@ -39,11 +39,20 @@
     if($_SERVER['REQUEST_METHOD'] === 'POST'){
       $id = $_POST['id'];
       $id = filter_var($id, FILTER_VALIDATE_INT);
+
       if($id){
+        $tipo = $_POST['tipo'];
+        if(validarTipoContenido($tipo)) {
+          // Compara lo que vamos a Eliminar
+          if($tipo === 'vendedor'){
+            $vendedor = Vendedor::findOne($id);
+            $vendedor->deleteOne();
 
-        $propiedad = Propiedad::findOne($id);
-
-        $propiedad->deleteOne();
+          }else if($tipo === 'propiedad'){
+            $propiedad = Propiedad::findOne($id);
+            $propiedad->deleteOne();
+          }
+        }
       }
     }
 
@@ -68,48 +77,82 @@
       case '3':
         echo '<p class="alerta exito">Propiedad eliminada correctamente</p>';
         break;
+      case '4':
+        echo '<p class="alerta exito">Vendedor eliminado correctamente</p>';
+        break;
       default:
         echo '';
         break;
     }
   ?>
-</main>
 
-<a href="/admin/propiedades/crear.php" class="boton boton-verde">Nueva Propiedad</a>
+  <a href="/admin/propiedades/crear.php" class="boton boton-verde">Nueva Propiedad</a>
 
-<table class="propiedades">
-  <thead>
-    <tr>
-      <th>ID</th>
-      <th>Titulo</th>
-      <th>Imagen</th>
-      <th>Precio</th>
-      <th>Acciones</th>
-    </tr>
-  </thead>
+  <h2>Propiedades</h2>
 
-  <tbody><!-- Mostrar los resultados de la BBDD en la tabla -->
+  <table class="propiedades">
+    <thead>
+      <tr>
+        <th>ID</th>
+        <th>Titulo</th>
+        <th>Imagen</th>
+        <th>Precio</th>
+        <th>Acciones</th>
+      </tr>
+    </thead>
+    
+    <tbody><!-- Mostrar los resultados de la BBDD en la tabla -->
     <?php foreach($propiedades as $propiedad):?>
       <tr>
         <td><?php echo $propiedad->id; ?></td>
         <td><?php echo $propiedad->titulo; ?></td>
         <td><img src="/imagenes/<?php echo $propiedad->imagen; ?>" alt="Imagen Propiedad" class="imagen-tabla"></td>
-        <td><?php echo $propiedad->precio; ?></td>
+        <td><?php echo $propiedad->precio; ?> €</td>
         <td>
           <a href="/admin/propiedades/actualizar.php?id=<?php echo $propiedad->id; ?>" class="boton-amarillo-block">Actualizar</a>
           <form method="POST" class="w-100">
             <input type="hidden" name="id" value="<?php echo $propiedad->id; ?>"/>
+            <input type="hidden" name="tipo" value="propiedad"/>
             <input type="submit" class="boton-rojo-block" value="Eliminar" />
           </form>
         </td>
       </tr>
-    <?php endforeach; ?>
-  </tbody>
-</table>
+      <?php endforeach; ?>
+    </tbody>
+  </table>
+
+  <h2>Vendedores</h2>
+
+  <table class="propiedades">
+    <thead>
+      <tr>
+        <th>ID</th>
+        <th>Nombre</th>
+        <th>Teléfono</th>
+        <th>Acciones</th>
+      </tr>
+    </thead>
+    
+    <tbody><!-- Mostrar los resultados de la BBDD en la tabla -->
+    <?php foreach($vendedores as $vendedor):?>
+      <tr>
+        <td><?php echo $vendedor->id; ?></td>
+        <td><?php echo $vendedor->nombre . " " . $vendedor->apellido; ?></td>
+        <td><?php echo $vendedor->telefono; ?></td>
+        <td>
+          <a href="/admin/vendedores/actualizar.php?id=<?php echo $vendedor->id; ?>" class="boton-amarillo-block">Actualizar</a>
+          <form method="POST" class="w-100">
+            <input type="hidden" name="id" value="<?php echo $vendedor->id; ?>"/>
+            <input type="hidden" name="tipo" value="vendedor"/>
+            <input type="submit" class="boton-rojo-block" value="Eliminar" />
+          </form>
+        </td>
+      </tr>
+      <?php endforeach; ?>
+    </tbody>
+  </table>
+</main>
 
 <?php
-  // Cerrar la conexión a la BBDD
-  mysqli_close($db);
-
   incluirTemplate('footer');
 ?>
